# Mechanic_App/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanics_API"
  description: "This is an API to manage a Mechanic shop database"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /service_tickets:
    post:
      tags:
        - ServiceTickets
      summary: "This creates a new service ticket base"
      description: "This endpoint creates a service ticket. Remeber to add the customer id because it is required."        
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Information required to create a service ticket"
          schema:
            $ref: "#/definitions/CreateServicePayload"
      responses:
        201:
          description: "Service ticket created sucessfully"
          schema:
            $ref: "#/definitions/CreateServiceResponse"
          examples:
            application/json:
              id: 1
              customer_id: 3
              vin: "1HGCM82633A004352"
              service_date: "2025-01-01"
              service_desc: "Oil change and tire rotation"
              mechanic_services: [1, 4]


    get:
      tags:
        - ServiceTickets
      summary: "Retrieves all the service tickets"
      description: "This endpoint retrieves all the service tickets that are currently in the database"
      responses: 
        200:
          description: "Retrieve service tickets successfully"
          schema:
            $ref: "#/definitions/RetrieveAllServiceTickets"
  /service_tickets/my-tickets:
    get:
      tags:
        - ServiceTickets
      summary: "retrieves customer ticket"
      description: "This endpoint retrieves the ticket based on the customer. It uses authentication so you must be logged in to access your ticket"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully retrieves customer's ticket"
          schema:
            $ref: "#/definitions/RetrieveCustomerTicket"
  /service_tickets/{ticket_id}/assign-mechanic/{mechanic_id}:
    put:
      tags:
        - ServiceTickets
      summary: "Assigns mechanic to an existing service ticket"
      description: "Adds a mechanic to an existing service ticket."
      parameters:
        - in: "path"
          name: "ticket_id"
          required: true
          type: "integer"
          description: "ID of the service ticket"

        - in: path
          name: "mechanic_id"
          required: true
          type: "integer"
          description: "ID of the mechanic you want to assign"
      responses:
        201: 
          description: "Mechanic assigned successfully"
          schema: 
            $ref: "#/definitions/AssignMechanicResponse"
          examples:
            application/json:
              Message: "Mechanic successfully added"
  /service_tickets/{ticket_id}/remove-mechanic/{mechanic_id}:
    put:
      tags:
        - ServiceTickets
      summary: "Remove mechanic from an existing service ticket"
      description: "Removes a mechanic from an existing service ticket."
      parameters:
        - in: "path"
          name: "ticket_id"
          required: true
          type: "integer"
          description: "ID of the service ticket"

        - in: path
          name: "mechanic_id"
          required: true
          type: "integer"
          description: "ID of the mechanic you want to remove"
      responses:
        201: 
          description: "Mechanic remove successfully"
          schema: 
            $ref: "#/definitions/RemoveMechanicResponse"
          examples:
            application/json:
              Message: "Mechanic successfully remove"
  /service_tickets/{ticket_id}/add-part/{inventory_id}:
    post:
      tags:
        - ServiceTickets
      summary: "Add parts to a service ticket"
      description: "This endpoint takes the part_id and add it to the ticket_id given."
      parameters:
        - in: "path"
          name: "ticket_id"
          required: true
          type: "integer"
          description: "ID of the service ticket"

        - in: path
          name: "inventory_id"
          required: true
          type: "integer"
          description: "ID of the part that you want to add to the service ticket"

        - in: "body"
          name: "body"
          required: true
          description: "Quantity of the part being added"
          schema:
            $ref: "#/definitions/AddPartPayload"
      responses:
        200: 
          description: "Successfully added part to service ticket"
          schema: 
            $ref: "#/definitions/AddPartResponse"
          examples:
            application/json:
              Message: "Successfully added part to the service ticket"

#Inventory
...
definitions:
...
  CreateServicePayload:
    type: "object"
    properties:
      customer_id:
        type: "integer"
      vin:
        type: "string"
      service_date:
        type: "string"
        format: "date"
      service_desc: 
        type: "string"
      mechanic_services:
        type: "array"
        items:
          type: "integer"

    required:
      - customer_id
      - vin
      - service_date
      - service_desc
  CreateServiceResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      customer_id:
        type: "integer"
      vin:
        type: "string"
        maxLength: 17
      service_date:
        type: "string"
        format: "date"
      service_desc: 
        type: "string"
      mechanic_services:
        type: "array"
        items:
          type: "integer"
  RetrieveAllServiceTickets:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        customer_id:
          type: "string"
        vin:
          type: "string"
        service_date:
          type: "string"
          format: "date"
        service_desc: 
          type: "string"
        mechanic_services:
          type: "array"
          items:
            type: "integer"
  AssignMechanicResponse:
    type: "object"
    properties:
      message: #You don't have to rewrite all the payload content it will be unecessary because we are only adding mechanic and techinally the ticket_id represents the whole service ticket
        type: "string"
  RemoveMechanicResponse:
    type: "object"
    properties:
      message: #You don't have to rewrite all the payload content it will be unecessary because we are only removing mechanic and techinally the ticket_id represents the whole service ticket
        type: "string"
  RetrieveCustomerTicket:
    type: "object"
    properties:
      message: #You don't have to rewrite all the payload content it will be unecessary because we are only removing mechanic and techinally the ticket_id represents the whole service ticket
        type: "string"
  AddPartPayload:
    type: object
    properties:
      quantity_used:
        type: "integer"
        description: "How many of this part were used"
    required:
      - quantity_used
  AddPartResponse:
    type: "object"
    properties:
      message: #You don't have to rewrite all the payload content it will be unecessary because we are only removing mechanic and techinally the ticket_id represents the whole service ticket
        type: "string"

#Inventory
...